requirements.txt
Требуются библиотеки:
sqlite3
PyQt
sys
